refactor: move Redis-related files to the redis package and add logging to EmbeddedRedisConfig
