+++ gateway/src/main/java/com/oing/config/redis/EmbeddedRedisConfig.java
@@ -1,12 +1,14 @@
-package com.oing.config;
+package com.oing.config.redis;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import redis.embedded.RedisServer;
 
+@Slf4j
 @Configuration
 @Profile("local")
 public class EmbeddedRedisConfig {
@@ -23,7 +25,9 @@ public class EmbeddedRedisConfig {
                     .port(port)
                     .setting("maxmemory 256M")
                     .build();
+            log.info("Embedded Redis Server Build");
             redisServer.start();
+            log.info("Embedded Redis Server Start");
         } catch (Exception e) {
             e.printStackTrace();
         }
