+++ gateway/src/main/java/com/oing/config/redis/RedisConfig.java
@@ -1,19 +1,22 @@
-package com.oing.config;
+package com.oing.config.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.data.redis.connection.RedisClusterConfiguration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceClientConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
-@Profile("local")
+import java.util.Collections;
+
+@Profile({"prod", "dev"})
 @Configuration
-public class RedisLocalConfig {
+public class RedisConfig {
 
     @Value("${spring.data.redis.host}")
     private String redisHost;
@@ -23,7 +26,8 @@ public class RedisLocalConfig {
 
     @Bean
     public RedisConnectionFactory redisConnectionFactory() {
-        return new LettuceConnectionFactory(new RedisStandaloneConfiguration(redisHost, redisPort));
+        LettuceClientConfiguration configuration = LettuceClientConfiguration.builder().useSsl().build();
+        return new LettuceConnectionFactory(new RedisClusterConfiguration(Collections.singleton(redisHost + ":" + redisPort)), configuration);
     }
 
     @Bean
